split votchallenge.rs into multiple files
